+++ apps/fabric-website/src/root.tsx
@@ -1,9 +1,11 @@
+import 'react-app-polyfill/ie11';
 import { registerIcons } from 'office-ui-fabric-react';
 import { initializeFileTypeIcons } from '@uifabric/file-type-icons';
 import { createSite } from './utilities/createSite';
 import * as platformPickerStyles from '@uifabric/example-app-base/lib/components/PlatformPicker/PlatformPicker.module.scss';
 import { SiteDefinition } from './SiteDefinition/index';
 import { HomePage } from './pages/HomePage/HomePage';
+import { NotFoundPage } from './pages/NotFoundPage/NotFoundPage';
 import { AndroidLogo, AppleLogo, WebLogo } from './utilities/index';
 
 // TODO: handle redirects
@@ -24,4 +26,4 @@ registerIcons({
   }
 });
 
-createSite(SiteDefinition, HomePage);
+createSite(SiteDefinition, [NotFoundPage, HomePage]);