+++ packages/experiments/src/components/Chiclet/ChicletPage.tsx
@@ -5,6 +5,8 @@ import { ChicletBasicExample } from './examples/Chiclet.Basic.Example';
 const ChicletBasicExampleCode = require('!raw-loader!@uifabric/experiments/src/components/Chiclet/examples/Chiclet.Basic.Example.tsx') as string;
 import { ChicletBreadcrumbExample } from './examples/Chiclet.Breadcrumb.Example';
 const ChicletBreadcrumbExampleCode = require('!raw-loader!@uifabric/experiments/src/components/Chiclet/examples/Chiclet.Breadcrumb.Example.tsx') as string;
+import { ChicletFooterExample } from './examples/Chiclet.Footer.Example';
+const ChicletFooterExampleCode = require('!raw-loader!@uifabric/experiments/src/components/Chiclet/examples/Chiclet.Footer.Example.tsx') as string;
 
 export class ChicletPage extends React.Component<IComponentDemoPageProps, {}> {
   public render(): JSX.Element {
@@ -19,6 +21,9 @@ export class ChicletPage extends React.Component<IComponentDemoPageProps, {}> {
             </ExampleCard>
             <ExampleCard title="Medium-sized Chiclet with Breadcrumb" isOptIn={true} code={ChicletBreadcrumbExampleCode}>
               <ChicletBreadcrumbExample />
+            </ExampleCard>
+            <ExampleCard title="Medium-sized Chiclet with Footer" isOptIn={true} code={ChicletFooterExampleCode}>
+              <ChicletFooterExample />
             </ExampleCard>
           </div>
         }