+++ packages/office-ui-fabric-react/src/components/ComboBox/ComboBox.tsx
@@ -25,7 +25,8 @@ import { getClassNames, getComboBoxOptionClassNames, IComboBoxClassNames } from 
 import { IComboBoxOption, IComboBoxOptionStyles, IComboBoxProps } from './ComboBox.types';
 import { KeytipData } from '../../KeytipData';
 import { Label } from '../../Label';
-import { SelectableOptionMenuItemType } from '../../utilities/selectableOption/SelectableOption.types';
+import { SelectableOptionMenuItemType, getAllSelectedOptions } from '../../utilities/selectableOption/index';
+import { BaseButton, Button } from '../Button/index';
 
 export interface IComboBoxState {
   /** The open state */
@@ -203,6 +204,15 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
       currentPendingValue: undefined,
       currentPendingValueValidIndexOnHover: HoverStatus.default
     };
+  }
+
+  /**
+   * All selected options
+   */
+  public get selectedOptions(): IComboBoxOption[] {
+    const { currentOptions, selectedIndices } = this.state;
+
+    return getAllSelectedOptions(currentOptions, selectedIndices!);
   }
 
   public componentDidMount(): void {
@@ -410,6 +420,7 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
                 data-is-focusable={false}
                 tabIndex={-1}
                 onClick={this._onComboBoxClick}
+                onBlur={this._onBlur}
                 iconProps={buttonIconProps}
                 disabled={disabled}
                 checked={isOpen}
@@ -830,7 +841,11 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
    * @param index - the index to set (or the index to set from if a search direction is provided)
    * @param searchDirection - the direction to search along the options from the given index
    */
-  private _setSelectedIndex(index: number, submitPendingValueEvent: any, searchDirection: SearchDirection = SearchDirection.none): void {
+  private _setSelectedIndex(
+    index: number,
+    submitPendingValueEvent: React.SyntheticEvent<any>,
+    searchDirection: SearchDirection = SearchDirection.none
+  ): void {
     const { onChange, onChanged, onPendingValueChanged } = this.props;
     const { currentOptions } = this.state;
     let { selectedIndices } = this.state;
@@ -867,24 +882,28 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
         selectedIndices[0] = index;
       }
 
-      // Set the selected option
-      this.setState({
-        selectedIndices: selectedIndices
-      });
+      submitPendingValueEvent.persist();
+      // Call onChange after state is updated
+      this.setState(
+        {
+          selectedIndices: selectedIndices
+        },
+        () => {
+          // If ComboBox value is changed, revert preview first
+          if (this._hasPendingValue && onPendingValueChanged) {
+            onPendingValueChanged();
+            this._hasPendingValue = false;
+          }
 
-      // If ComboBox value is changed, revert preview first
-      if (this._hasPendingValue && onPendingValueChanged) {
-        onPendingValueChanged();
-        this._hasPendingValue = false;
-      }
+          if (onChange) {
+            onChange(submitPendingValueEvent, option, index, undefined);
+          }
 
-      if (onChange) {
-        onChange(submitPendingValueEvent, option, index, undefined);
-      }
-
-      if (onChanged) {
-        onChanged(option, index, undefined, submitPendingValueEvent);
-      }
+          if (onChanged) {
+            onChanged(option, index, undefined, submitPendingValueEvent);
+          }
+        }
+      );
     }
 
     // clear all of the pending info
@@ -901,7 +920,12 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
     }
 
     if (!this.state.focused) {
-      this.setState({ focused: true });
+      const state: Pick<IComboBoxState, 'focused'> | Pick<IComboBoxState, 'focused' | 'isOpen'> = { focused: true };
+
+      if (this.props.openOnKeyboardFocus && !this.state.isOpen && !this.props.disabled) {
+        (state as Pick<IComboBoxState, 'focused' | 'isOpen'>).isOpen = true;
+      }
+      this.setState(state);
     }
   };
 
@@ -940,7 +964,9 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
    * OnBlur handler. Set the focused state to false
    * and submit any pending value
    */
-  private _onBlur = (event: React.FocusEvent<HTMLInputElement>): void => {
+  private _onBlur = (
+    event: React.FocusEvent<HTMLInputElement | Autofill | HTMLAnchorElement | HTMLButtonElement | HTMLDivElement | BaseButton | Button>
+  ): void => {
     // Do nothing if the blur is coming from something
     // inside the comboBox root or the comboBox menu since
     // it we are not really bluring from the whole comboBox
@@ -1967,7 +1993,13 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
     const { comboBoxOptionStyles: customStylesForAllOptions } = this.props;
     const { styles: customStylesForCurrentOption } = item;
 
-    return getOptionStyles(this.props.theme!, customStylesForAllOptions, customStylesForCurrentOption, this._isPendingOption(item));
+    return getOptionStyles(
+      this.props.theme!,
+      customStylesForAllOptions,
+      customStylesForCurrentOption,
+      this._isPendingOption(item),
+      item.hidden
+    );
   }
 
   /**