+++ packages/office-ui-fabric-react/src/components/Dropdown/Dropdown.types.ts
@@ -10,10 +10,17 @@ import { IPanelStyleProps } from '../Panel/Panel.types';
 
 export { SelectableOptionMenuItemType as DropdownMenuItemType } from '../../utilities/selectableOption/SelectableOption.types';
 
+export { ResponsiveMode }; // Exported because the type is an optional prop and not exported otherwise.
+
 /**
  * {@docCategory Dropdown}
  */
 export interface IDropdown {
+  /**
+   * All selected options
+   */
+  readonly selectedOptions: IDropdownOption[];
+
   focus: (shouldOpenOnFocus?: boolean) => void;
 }
 
@@ -229,6 +236,9 @@ export interface IDropdownStyles {
 
   /** Style for a dropdown item when it is both selected and disabled. */
   dropdownItemSelectedAndDisabled: IStyle;
+
+  /** Style for a dropdown item when it is hidden */
+  dropdownItemHidden: IStyle;
 
   /**
    * Refers to the text element that renders the actual dropdown item/option text. This would be wrapped by the element