+++ packages/api-docs/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @uifabric/api-docs
 
-This log was last generated on Sat, 04 May 2019 00:01:53 GMT and should not be manually modified.
+This log was last generated on Tue, 14 May 2019 07:50:30 GMT and should not be manually modified.
+
+## 1.1.1
+Tue, 14 May 2019 07:50:30 GMT
+
+### Patches
+
+- Update Fabric assets link
+
+## 1.1.0
+Tue, 14 May 2019 01:56:43 GMT
+
+### Minor changes
+
+- PageJsonGenerator: handle type aliases.
 
 ## 1.0.1
 Sat, 04 May 2019 00:01:53 GMT